requirements.txt
aidial-sdk==0.27.0
aidial-client==0.3.0
mcp==1.20.0
pydantic==2.12.3
faiss-cpu==1.12.0
sentence-transformers==5.1.2
beautifulsoup4==4.14.2
pdfplumber==0.11.7
numpy==2.3.4
pandas==2.3.3
tabulate==0.9.0
langchain==1.0.3
langchain-text-splitters==1.0.0